test(llc): migrate client sync tests to stream_chat_test
